functions/CoverProcessor/requirements_minimal.txt
boto3
google-cloud-vision